changed the architecture to be based on controllers, added babel config to be able to write ES6, replaced the prices endpoint to use coingecko
