+++ .babelrc
@@ -1,0 +1,9 @@
+{
+  "presets": [
+    "@babel/preset-env",
+  ],
+  "plugins": [
+    ["@babel/plugin-transform-runtime"],
+    ["@babel/plugin-proposal-export-namespace-from"],
+  ]
+}
