+++ .gitignore
@@ -1,2 +1,4 @@
 node_modules
 .env
+lib
+.editorconfig
