--- mybit-go/index.js
@@ -1,7 +1,0 @@
-const app = require('express')()
-
-app.get('*', (req, res) => {
-  res.send('MyBit Go API Endpoint');
-})
-
-app.listen(process.env.PORT || 8081);
