+++ package.json
@@ -1,21 +1,20 @@
 {
   "name": "MyBit.tech",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "lib/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start:mybit": "node mybit/index.js",
-    "start:mybit-go": "node mybit-go/index.js",
-    "start:token-sale": "node token-sale/index.js",
+    "start-dev": "NODE_ENV=development babel-watch --watch src src/index.js",
+    "compile": "rimraf lib/ && babel -d lib/ src/",
     "start:version": "node version/index.js",
-    "now-start": "node token-sale/index.js",
-    "start-dev": "NODE_ENV=development node token-sale/index.js"
+    "now-start": "npm run compile && node lib/index.js"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@babel/runtime": "^7.3.1",
     "cors": "^2.8.5",
     "dayjs": "^1.7.8",
     "dotenv": "^6.2.0",
@@ -23,5 +22,14 @@
     "express": "^4.16.4",
     "isomorphic-unfetch": "^3.0.0",
     "web3": "1.0.0-beta.35"
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.2.2",
+    "@babel/plugin-proposal-export-namespace-from": "^7.2.0",
+    "@babel/plugin-transform-runtime": "^7.2.0",
+    "@babel/preset-env": "^7.3.1",
+    "babel-watch": "^7.0.0",
+    "rimraf": "^2.6.3"
   }
 }
