+++ src/constants/gasPrice.js
@@ -1,0 +1,1 @@
+export const GAS_PRICE = 5000000000;
