+++ src/constants/index.js
@@ -1,0 +1,4 @@
+export * from './list_tokens_prices'
+export * from './gasPrice';
+export { AssetCollateral } from './contracts/AssetCollateralContract';
+export { TokenSale } from './contracts/TokenSaleContract';
