+++ src/constants/list_tokens_prices.js
@@ -1,0 +1,4 @@
+export const LIST_TOKENS_PRICES = [
+  'ethereum',
+  'mybit-token',
+];
