+++ src/controllers/index.js
@@ -1,0 +1,3 @@
+export * as CollateralController from './collateralController';
+export * as PricesController from './pricesController';
+export * as TokenSaleController from './tokenSaleController';
