--- version/version
@@ -1,1 +1,0 @@
-No version was provided
